perf(core): cut frames before squashing, and pin the channel-weight dtype
Two small fixes in the greyscale path:

`DummyCore.forward` squashed all 150 frames and then discarded the first 30. The squash is
pointwise in time, so cutting first is exactly equivalent and avoids building an intermediate
25% larger than needed on every batch.

`WeightedChannelSumLayer` built its parameter with `torch.tensor(init_channel_weights)`, so
integer weights - and `configs/model/linear_nonlinear_poisson.yaml` invites exactly that with
`[0.5, 0.5]` as its example, next to which `[1, 0]` reads as natural - produced an int64
parameter. `.float()` and `.half()` skip non-floating parameters, and `trainable=True` would
raise on requires_grad, so the dtype is now explicit.

+++ openretina/modules/layers/reducers.py
@@ -12,7 +12,11 @@ class WeightedChannelSumLayer(nn.Module):
         super().__init__()
 
         # add the channel weights
-        self.channel_weights = nn.Parameter(torch.tensor(init_channel_weights), requires_grad=trainable)
+        # dtype is explicit: integer weights such as [1, 0] would otherwise make an int64
+        # parameter, which .float()/.half() skip and which cannot carry gradients.
+        self.channel_weights = nn.Parameter(
+            torch.tensor(init_channel_weights, dtype=torch.float32), requires_grad=trainable
+        )
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         """If the input is not already single-channel (i.e. greyscale), take a weighted sum over channels ."""
